Add tests for different decimal scenarios (#123)
This PR adds tests to cover different decimal scenarios. 

It also workarounds an ongoing issue on small configurables (u8,
booleans): https://github.com/FuelLabs/fuels-rs/issues/1265. I will open
an issue as soon as this PR is merged to go back to u8 configurables
when the fuels-rs bug is fixed.

No major changes in the sway code. Decimal conversion is working fine, I
only had to add the test cases.

+++ packages/fungible-token/bridge-fungible-token/tests/functions/bridge/mod.rs
@@ -520,7 +520,7 @@ mod success {
 
         let response = bridged_token_decimals(&contract).await;
 
-        assert_eq!(response, BRIDGED_TOKEN_DECIMALS)
+        assert_eq!(u64::from(response), BRIDGED_TOKEN_DECIMALS)
     }
 
     #[tokio::test]
